requirements.txt
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
idna==3.7
prometheus_client==0.20.0
pycparser==2.22
pycryptodome==3.20.0
PyYAML==6.0.1
requests==2.32.3
tinytuya==1.14.0
tuyapower==0.2.0
urllib3==2.2.2
